Move the rbac user cluster controller
